Rename Protocol class to MiIOProtocol
