+++ miio/plug_cli.py
@@ -7,7 +7,7 @@ import click
 
 import miio  # noqa: E402
 from miio.click_common import ExceptionHandlerGroup, validate_ip, validate_token
-from miio.protocol import Protocol
+from miio.miioprotocol import MiIOProtocol
 
 _LOGGER = logging.getLogger(__name__)
 pass_dev = click.make_pass_decorator(miio.ChuangmiPlug)
@@ -46,7 +46,7 @@ def cli(ctx, ip: str, token: str, debug: int):
 @cli.command()
 def discover():
     """Search for plugs in the network."""
-    Protocol.discover()
+    MiIOProtocol.discover()
 
 
 @cli.command()
